Refactor into own file
Help to clean files and improve imports
